requirements.txt
absl-py==2.3.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aiosqlite==0.19.0
alembic==1.15.2
altair==5.5.0
annotated-types==0.7.0
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
asttokens==3.0.0
asyncio-mqtt==0.16.1
attrs==25.3.0
auth0-python==4.9.0
autograd==1.8.0
autograd-gamma==0.5.0
babel==2.17.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.12.2
black==23.11.0
blinker==1.9.0
Brotli==1.1.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.8
coloredlogs==15.0.1
contourpy==1.3.2
courlan==1.3.2
crewai==0.14.0
crewai-tools==0.0.6
cryptography==43.0.3
cycler==0.12.1
dataclasses-json==0.6.7
dateparser==1.2.2
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
docstring-parser==0.15
duckdb==1.3.2
durationpy==0.9
ecos==2.0.14
embedchain==0.1.102
environs==9.5.0
et_xmlfile==2.0.0
executing==2.2.0
fastapi==0.104.1
feedparser==6.0.11
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.4
formulaic==1.2.0
frozenlist==1.6.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.4.0
google-api-core==2.24.2
google-api-python-client==2.110.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0
google-cloud-aiplatform==1.90.0
google-cloud-bigquery==3.30.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-generativeai==0.3.2
google-resumable-media==2.7.2
google_search_results==2.4.2
googleapis-common-protos==1.70.0
gptcache==0.1.44
greenlet==3.2.1
grpc-google-iam-v1==0.14.2
grpcio==1.58.0
grpcio-status==1.58.0
h11==0.16.0
htmldate==1.9.3
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.25.2
huggingface-hub==0.30.2
humanfriendly==10.0
icalendar==5.0.11
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
iniconfig==2.1.0
instructor==0.5.2
interface-meta==1.3.0
ipython==9.1.0
ipython_pygments_lexers==1.1.1
isort==5.12.0
jax==0.4.38
jaxlib==0.4.38
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.1
json5==0.12.0
json_repair==0.42.0
jsonpatch==1.33
jsonpickle==4.0.5
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jusText==3.0.2
kiwisolver==1.4.8
kubernetes==32.0.1
lancedb==0.5.7
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.53
langchain-openai==0.0.5
langchain-text-splitters==0.0.2
langsmith==0.1.147
lifelines==0.30.0
litellm==1.40.8
lxml==5.4.0
lxml_html_clean==0.4.2
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.21
minio==7.2.16
ml_dtypes==0.5.1
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
msal==1.25.0
msgraph-core==0.2.2
multidict==6.4.3
mutagen==1.47.0
mypy_extensions==1.1.0
narwhals==2.1.2
networkx==3.4.2
nodeenv==1.9.1
numexpr==2.11.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.21.1
openai==1.76.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.32.1
opentelemetry-exporter-otlp-proto-common==1.32.1
opentelemetry-exporter-otlp-proto-grpc==1.32.1
opentelemetry-exporter-otlp-proto-http==1.32.1
opentelemetry-instrumentation==0.53b1
opentelemetry-instrumentation-asgi==0.53b1
opentelemetry-instrumentation-fastapi==0.53b1
opentelemetry-proto==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
opentelemetry-util-http==0.53b1
opt_einsum==3.4.0
orjson==3.10.16
osqp==0.6.7.post3
overrides==7.7.0
packaging==23.2
paho-mqtt==2.1.0
pandas==2.3.1
parso==0.8.4
pathspec==0.12.1
pdfminer.six==20250327
pdfplumber==0.11.6
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.5.0
polars==1.32.3
posthog==3.25.0
prometheus-client==0.19.0
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==4.25.8
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pulsar-client==3.6.1
pure_eval==0.2.3
py==1.11.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pycryptodomex==3.22.0
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pylance==0.9.18
pymilvus==2.3.4
pyparsing==3.2.3
pypdf==3.17.4
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyright==1.1.400
pysbd==0.3.4
pytest==7.4.3
pytest-asyncio==0.21.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.9
pytube==15.0.0
pytz==2025.2
pyvis==0.3.2
PyYAML==6.0.2
qdldl==0.1.7.post5
RapidFuzz==3.14.0
ratelimiter==1.2.0.post0
referencing==0.36.2
regex==2025.7.34
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retry==0.9.2
rich==13.7.0
rpds-py==0.24.0
rsa==4.9.1
safetensors==0.4.1
schema==0.7.7
scikit-learn==1.7.1
scikit-survival==0.25.0
scipy==1.16.0
semver==3.0.4
sentencepiece==0.2.0
sgmllib3k==1.0.0
shapely==2.1.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.2
soupsieve==2.7
SQLAlchemy==2.0.27
stack-data==0.6.3
starlette==0.27.0
streamlit==1.48.1
sympy==1.13.3
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.5.2
tld==0.13.1
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
tomli_w==1.2.0
torch==2.1.0
tornado==6.5.2
tqdm==4.67.1
trafilatura==2.0.0
traitlets==5.14.3
typer==0.9.4
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.1.0
uv==0.6.16
uvicorn==0.24.0
uvloop==0.21.0
# Editable Git install with no remote (valetrisk-poc==0.0.1)
-e /Users/pinaz.patel/Downloads/valetrisk-poc
watchfiles==1.0.5
wcwidth==0.2.13
websocket-client==1.8.0
websockets==12.0
wrapt==1.17.2
xgboost==3.0.4
yarl==1.20.0
youtube-transcript-api==0.6.3
yt-dlp==2023.12.30
zipp==3.21.0
zstandard==0.23.0
